+++ src/components/Faculty/CardStyle.module.css
@@ -1,14 +1,3 @@
-.facultyContainer {
-  margin: 50px 4rem;
-}
-
-.facultyHeading {
-  background: #fff;
-  font-weight: 500;
-  text-align: center;
-  margin-bottom: 20px;
-}
-
 .facultyCard {
   border-radius: 3%;
   width: 100%;
@@ -42,7 +31,3 @@
   font-size: large;
   word-break: break-all;
 }
-/* .SalmanImagePosition {
-  object-fit: none;
-  object-position: 70% 0%;
-} */
