+++ src/components/Faculty/FacultyStyle.module.css
@@ -1,0 +1,16 @@
+.facultyContainer {
+  margin: 50px 4rem;
+}
+
+.facultyHeading {
+  background: #fff;
+  font-weight: 500;
+  text-align: center;
+  margin-bottom: 20px;
+}
+
+@media screen and (max-width: 786px) {
+  .facultyContainer {
+    margin: 50px 1.5rem;
+  }
+}
